requirements.txt
streamlit
torch
torchvision
opencv-python
pillow
matplotlib
numpy
